Revise data import for Axios and Vue Query to improve code splitting
