+++ services/agora/src/boot/axios.ts
@@ -1,5 +1,6 @@
 import { defineBoot } from "#q-app/wrappers";
-import axios, { type AxiosInstance } from "axios";
+import type { AxiosInstance } from "axios";
+import { axiosInstance, api } from "src/utils/api/client";
 
 declare module "vue" {
   interface ComponentCustomProperties {
@@ -8,18 +9,10 @@ declare module "vue" {
   }
 }
 
-// Be careful when using SSR for cross-request state pollution
-// due to creating a Singleton instance here;
-// If any client changes this (global) instance, it might be a
-// good idea to move this instance creation inside of the
-// "export default () => {}" function below (which runs individually
-// for each client)
-const api = axios.create({ baseURL: process.env.VITE_API_BASE_URL });
-
 export default defineBoot(({ app }) => {
   // for use inside Vue files (Options API) through this.$axios and this.$api
 
-  app.config.globalProperties.$axios = axios;
+  app.config.globalProperties.$axios = axiosInstance;
   // ^ ^ ^ this will allow you to use this.$axios (for Vue Options API form)
   //       so you won't necessarily have to import axios in each vue file
 
@@ -27,5 +20,3 @@ export default defineBoot(({ app }) => {
   // ^ ^ ^ this will allow you to use this.$api (for Vue Options API form)
   //       so you can easily perform requests against your app's API
 });
-
-export { axios, api };
