+++ services/agora/src/utils/api/administrator/organization.ts
@@ -1,4 +1,4 @@
-import { api } from "boot/axios";
+import { api } from "../client";
 import type {
   ApiV1AdministratorOrganizationAddUserOrganizationMappingPostRequest,
   ApiV1AdministratorOrganizationCreateOrganizationPostRequest,
