+++ services/agora/src/utils/api/client.ts
@@ -1,0 +1,5 @@
+import axios from "axios";
+
+// Create axios instances that can be imported
+export const axiosInstance = axios;
+export const api = axios.create({ baseURL: process.env.VITE_API_BASE_URL });
