+++ services/agora/src/utils/api/topic.ts
@@ -3,7 +3,7 @@ import type {
   ApiV1TopicGetFollowedPost200Response,
 } from "src/api";
 import { DefaultApiAxiosParamCreator, DefaultApiFactory } from "src/api";
-import { api } from "src/boot/axios";
+import { api } from "./client";
 import type { AxiosErrorResponse, AxiosSuccessResponse } from "./common";
 import { useCommonApi } from "./common";
 import type { GetAllTopicsResponse } from "src/shared/types/dto";
